+++ client/internal/wgproxy/ebpf/bpf_bpfeb.go
@@ -2,7 +2,7 @@
 //go:build arm64be || armbe || mips || mips64 || mips64p32 || ppc64 || s390 || s390x || sparc || sparc64
 // +build arm64be armbe mips mips64 mips64p32 ppc64 s390 s390x sparc sparc64
 
-package wgproxy
+package ebpf
 
 import (
 	"bytes"
