+++ client/internal/wgproxy/ebpf/bpf_bpfeb.o
