+++ client/internal/wgproxy/ebpf/bpf_bpfel.go
@@ -2,7 +2,7 @@
 //go:build 386 || amd64 || amd64p32 || arm || arm64 || mips64le || mips64p32le || mipsle || ppc64le || riscv64
 // +build 386 amd64 amd64p32 arm arm64 mips64le mips64p32le mipsle ppc64le riscv64
 
-package wgproxy
+package ebpf
 
 import (
 	"bytes"
