+++ client/internal/wgproxy/ebpf/bpf_bpfel.o
