+++ client/internal/wgproxy/ebpf/loader_test.go
@@ -1,16 +1,16 @@
 //go:build linux
 
-package wgproxy
+package ebpf
 
 import (
 	"testing"
 )
 
 func Test_newEBPF(t *testing.T) {
-	ebpf := newEBPF()
-	err := ebpf.load(1234, 51892)
+	ebpf := NewEBPF()
+	err := ebpf.Load(1234, 51892)
 	defer func() {
-		_ = ebpf.free()
+		_ = ebpf.Free()
 	}()
 	if err != nil {
 		t.Errorf("%s", err)
